tx list encode, decode, validate, prepare
